Fixed displaying the security descriptor list in ntfsinfo
The displaying of the security descriptors in $Secure/$SDS was incomplete,
moreover the recent limitations for safety purposes in reading an
attribute in a single step made it to fail.
With this patch the security descriptors in use are accessed through
the $SII index and displayed ordered by their ids.